% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multiclass.R
\name{ACC}
\alias{ACC}
\title{Accuracy}
\usage{
ACC(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Defined as: mean(response == truth)
}
