% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{BER}
\alias{BER}
\title{Balanced error rate}
\usage{
BER(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Mean of misclassification error rates on all individual classes.
}
