% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_binaryclass.R
\name{F1}
\alias{F1}
\title{F1 measure}
\usage{
F1(truth, response, positive)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{positive}{positive class}
}
\description{
Defined as: 2 * tp/ (sum(truth == positive) + sum(response == positive))
}
