% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_binaryclass.R
\name{FDR}
\alias{FDR}
\title{False discovery rate}
\usage{
FDR(truth, response, positive)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{positive}{positive class}
}
\description{
Defined as: fp / (tp + fp)
}
