% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_binaryclass.R
\name{FN}
\alias{FN}
\title{False negatives}
\usage{
FN(truth, response, negative)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{negative}{negative class}
}
\description{
Sum of misclassified observations in the negative class. Also called misses.
}
