% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{FP}
\alias{FP}
\title{False positives}
\usage{
FP(truth, response, positive)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{positive}{positive class}
}
\description{
Sum of misclassified observations in the positive class. Also called false alarms.
}
