% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multiclass.R
\name{KAPPA}
\alias{KAPPA}
\title{Cohen's kappa}
\usage{
KAPPA(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Defined as: 1 - (1 - p0) / (1 - pe). With: p0 = 'observed frequency of
agreement' and pe = 'expected agremeent frequency under independence
}
