% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllMeasures.R
\name{listAllMeasures}
\alias{listAllMeasures}
\title{List all measures}
\usage{
listAllMeasures()
}
\value{
Dataframe with all available measures and the correspoding task
}
\description{
Lists all measures that are available in the package with their corresponding task.
}
\examples{
listAllMeasures()
}
