% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{Logloss}
\alias{Logloss}
\title{Logarithmic loss}
\usage{
Logloss(probabilities, truth)
}
\arguments{
\item{probabilities}{[numeric] vector of predicted probabilities}

\item{truth}{vector of true values}
}
\description{
Defined as: -mean(log(p_i)), where p_i is the predicted probability of the true 
class of observation i. Inspired by https://www.kaggle.com/wiki/MultiClassLogLoss.
}
