% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{MAPE}
\alias{MAPE}
\title{Mean absolute percentage error}
\usage{
MAPE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as the abs(truth_i - response_i) / truth_i. Won't work if any truth value is equal to zero. In this case the output will be NA.
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
MAPE(truth, response)
}
