% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{MEDAE}
\alias{MEDAE}
\title{Median of absolute errors}
\usage{
MEDAE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as: median(abs(response - truth)).
}
