% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{MMCE}
\alias{MMCE}
\title{Mean misclassification error}
\usage{
MMCE(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Defined as: mean(response != truth)
}
