% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{MSE}
\alias{MSE}
\title{Mean of squared errors}
\usage{
MSE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as: mean((response - truth)^2)
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
MSE(truth, response)
}
