% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multiclass.R
\name{multiclass.AU1P}
\alias{multiclass.AU1P}
\title{Weighted average 1 vs. 1 multiclass AUC}
\usage{
multiclass.AU1P(probabilities, truth)
}
\arguments{
\item{probabilities}{[numeric] matrix of predicted probabilities with columnnames of the classes}

\item{truth}{vector of true values}
}
\description{
Computes AUC of c(c - 1) binary classifiers while considering the a priori distribution of the classes. 
See Ferri et al.: https://www.math.ucdavis.edu/~saito/data/roc/ferri-class-perf-metrics.pdf.
}
