% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{multiclass.AUNP}
\alias{multiclass.AUNP}
\title{Weighted average 1 vs. rest multiclass AUC}
\usage{
multiclass.AUNP(probabilities, truth)
}
\arguments{
\item{probabilities}{[numeric] vector of predicted probabilities}

\item{truth}{vector of true values}
}
\description{
Computes the AUC treating a c-dimensional classifier as c two-dimensional classifiers, 
taking into account the prior probability of each class. 
See Ferri et al.: https://www.math.ucdavis.edu/~saito/data/roc/ferri-class-perf-metrics.pdf.
}
