% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{multiclass.Brier}
\alias{multiclass.Brier}
\title{Multiclass Brier score}
\usage{
multiclass.Brier(probabilities, truth)
}
\arguments{
\item{probabilities}{[numeric] vector of predicted probabilities}

\item{truth}{vector of true values}
}
\description{
Defined as: (1/n) sum_i sum_j (y_ij - p_ij)^2, where y_ij = 1 if observation i has class j (else 0), 
and p_ij is the predicted probability of observation i for class j. 
From http://docs.lib.noaa.gov/rescue/mwr/078/mwr-078-01-0001.pdf.
}
