% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{MultilabelACC}
\alias{MultilabelACC}
\title{Accuracy (multilabel)}
\usage{
MultilabelACC(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Averaged proportion of correctly predicted labels with respect to the total number of labels for each instance,
following the definition by Charte and Charte: https: / /journal.r-project.org / archive / 2015 - 2 / charte-charte.pdf.
Fractions where the denominator becomes 0 are replaced with 1 before computing the average across all instances.
}
