% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multilabel.R
\name{MultilabelF1}
\alias{MultilabelF1}
\title{F1 measure (multilabel)}
\usage{
MultilabelF1(truth, response)
}
\arguments{
\item{truth}{matrix of true values}

\item{response}{matrix of predicted values
n = 20
set.seed(122)
truth = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
response = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
MultilabelF1(truth, response)}
}
\description{
Harmonic mean of precision and recall on a per instance basis (Micro-F1), following the
definition by Montanes et al.: http: / /www.sciencedirect.com / science / article / pii / S0031320313004019.
Fractions where the denominator becomes 0 are replaced with 1 before computing the average across all instances.
}
