% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{MultilabelHamloss}
\alias{MultilabelHamloss}
\title{Hamming loss}
\usage{
MultilabelHamloss(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Proportion of labels that are predicted incorrectly, following the definition
by Charte and Charte: https://journal.r-project.org/archive/2015-2/charte-charte.pdf.
}
