% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multilabel.R
\name{MultilabelPPV}
\alias{MultilabelPPV}
\title{Positive predictive value (multilabel)}
\usage{
MultilabelPPV(truth, response)
}
\arguments{
\item{truth}{matrix of true values}

\item{response}{matrix of predicted values
n = 20
set.seed(122)
truth = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
response = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
MultilabelPPV(truth, response)}
}
\description{
Also called precision. Averaged ratio of correctly predicted labels for each instance,
following the definition by Charte and Charte: https: / /journal.r-project.org / archive / 2015 - 2 / charte-charte.pdf.
Fractions where the denominator becomes 0 are ignored in the average calculation.
}
