% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{MultilabelSubset01}
\alias{MultilabelSubset01}
\title{Subset-0-1 loss}
\usage{
MultilabelSubset01(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Proportion of observations where the complete multilabel set (all 0-1-labels) is predicted incorrectly,
following the definition by Charte and Charte: https://journal.r-project.org/archive/2015-2/charte-charte.pdf.
}
