% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{RAE}
\alias{RAE}
\title{Relative absolute error}
\usage{
RAE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as sum_of_absolute_errors / mean_absolute_deviation. Undefined for single instances and when every truth value is identical. In this case the output will be NA.
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
RAE(truth, response)
}
