% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{RMSE}
\alias{RMSE}
\title{Root mean squared error}
\usage{
RMSE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
The RMSE is aggregated as sqrt(mean(rmse.vals.on.test.sets^2))
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
RMSE(truth, response)
}
