% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{RMSLE}
\alias{RMSLE}
\title{Root mean squared logarithmic error}
\usage{
RMSLE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Definition taken from: https: / /www.kaggle.com / wiki / RootMeanSquaredLogarithmicError.
This  is mostly used for count data, note that all predicted and actual target values
must be greater or equal '-1' to compute the root mean squared logarithmic error.
}
\examples{
n = 20
set.seed(123)
truth = abs(rnorm(n))
response = abs(rnorm(n))
RMSLE(truth, response)
}
