% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{RSQ}
\alias{RSQ}
\title{Coefficient of determination}
\usage{
RSQ(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Also called R-squared, which is 1 - residual_sum_of_squares / total_sum_of_squares.
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
RSQ(truth, response)
}
