% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{SAE}
\alias{SAE}
\title{Sum of absolute errors}
\usage{
SAE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as: sum(abs(response - truth))"
}
