% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{TN}
\alias{TN}
\title{True negatives}
\usage{
TN(truth, response, negative)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{negative}{negative class}
}
\description{
Sum of correctly classified observations in the negative class. Also called correct rejections.
}
