% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{TNR}
\alias{TNR}
\title{True negative rate}
\usage{
TNR(truth, response, negative)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}

\item{negative}{negative class}
}
\description{
Percentage of correctly classified observations in the negative class. Also called specificity.
}
