% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{EXPVAR}
\alias{EXPVAR}
\title{Explained variance}
\usage{
EXPVAR(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Similar to RSQ (R-squared). Defined as explained_sum_of_squares / total_sum_of_squares.
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
EXPVAR(truth, response)
}
