% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multiclass.R
\name{MMCE}
\alias{MMCE}
\title{Mean misclassification error}
\usage{
MMCE(truth, response)
}
\arguments{
\item{truth}{vector of true values}

\item{response}{vector of predicted values}
}
\description{
Defined as: mean(response != truth)
}
\examples{
n = 20
set.seed(122)
truth = as.factor(sample(c(1,2,3), n, replace = TRUE))
response = as.factor(sample(c(1,2,3), n, replace = TRUE))
MMCE(truth, response)
}
