% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_multilabel.R
\name{MultilabelACC}
\alias{MultilabelACC}
\title{Accuracy (multilabel)}
\usage{
MultilabelACC(truth, response)
}
\arguments{
\item{truth}{matrix of true values}

\item{response}{matrix of predicted values
n = 20
set.seed(122)
truth = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
response = matrix(sample(c(0,1), 60, replace = TRUE), 20, 3)
MultilabelACC(truth, response)}
}
\description{
Averaged proportion of correctly predicted labels with respect to the total number of labels for each instance,
following the definition by Charte and Charte: https: / /journal.r-project.org / archive / 2015 - 2 / charte-charte.pdf.
Fractions where the denominator becomes 0 are replaced with 1 before computing the average across all instances.
}
