% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_regression.R
\name{SSE}
\alias{SSE}
\title{Sum of squared errors}
\usage{
SSE(truth, response)
}
\arguments{
\item{truth}{[numeric] vector of true values}

\item{response}{[numeric] vector of predicted values}
}
\description{
Defined as: sum((response - truth)^2)
}
\examples{
n = 20
set.seed(123)
truth = rnorm(n)
response = rnorm(n)
SSE(truth, response)
}
